get task detail
