+++ src/W2.Application.Contracts/Tasks/ITaskAppService.cs
@@ -23,5 +23,6 @@ namespace W2.Tasks
         Task<PagedResultDto<W2TasksDto>> DynamicDataByIdAsync(TaskDynamicDataInput input);
         Task<Dictionary<string, string>> handleDynamicData(TaskDynamicDataInput input);
         Task<List<DynamicDataDto>> GetDynamicRawData(TaskDynamicDataInput input);
+        Task<TaskDetailDto> GetDetailByWfId(string id);
     }
 }
